Rename example to MongoDB Realm
