+++ examples/mongoDbRealmExample/gradle.properties
