+++ examples/mongoDbRealmExample/src/main/AndroidManifest.xml
@@ -1,14 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="io.realm.examples.objectserver" >
+    package="com.mongodb.realm.example" >
 
     <application
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
-        android:name="io.realm.examples.objectserver.MyApplication"
+        android:name=".MyApplication"
         android:theme="@style/AppTheme" >
         <activity
-            android:name="io.realm.examples.objectserver.CounterActivity"
+            android:name=".CounterActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -16,7 +16,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="io.realm.examples.objectserver.LoginActivity"
+            android:name=".LoginActivity"
             android:label="Login" >
         </activity>
     </application>
