+++ examples/mongoDbRealmExample/src/main/res/drawable-xxhdpi/ic_exit_to_app_white_24dp.png
