+++ examples/mongoDbRealmExample/src/main/res/drawable-xxxhdpi/ic_exit_to_app_white_24dp.png
