+++ examples/mongoDbRealmExample/src/main/res/drawable/button_counter.xml
