+++ examples/mongoDbRealmExample/src/main/res/drawable/logo.png
