+++ examples/mongoDbRealmExample/src/main/res/layout/activity_counter.xml
