+++ examples/mongoDbRealmExample/src/main/res/layout/activity_login.xml
