+++ examples/mongoDbRealmExample/src/main/res/menu/menu_counter.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
       xmlns:app="http://schemas.android.com/apk/res-auto"
       xmlns:tools="http://schemas.android.com/tools"
-      tools:context=".CounterActivity">
+      tools:context="com.mongodb.realm.example.CounterActivity">
     <item
         android:id="@+id/action_logout"
         android:orderInCategory="100"
