+++ examples/mongoDbRealmExample/src/main/res/mipmap-xhdpi/ic_launcher.png
