+++ examples/mongoDbRealmExample/src/main/res/values-w820dp/dimens.xml
