+++ examples/mongoDbRealmExample/src/main/res/values/realm_colors.xml
