+++ examples/mongoDbRealmExample/src/main/res/values/styles.xml
